+++ src/main/java/me/meloni/SolarLogAPI/Interface/BasicUI/DayCustomizer.java
@@ -1,9 +1,9 @@
 package me.meloni.SolarLogAPI.Interface.BasicUI;
 
-import me.meloni.SolarLogAPI.Interface.DatePicker;
+import me.meloni.SolarLogAPI.Interface.Components.DatePicker;
 import me.meloni.SolarLogAPI.DataConversion.GetStartOf;
 import me.meloni.SolarLogAPI.SolarMap;
-import me.meloni.SolarLogAPI.Interface.Graph.DayView;
+import me.meloni.SolarLogAPI.Interface.Components.Graph.DayView;
 
 import javax.swing.*;
 import java.awt.*;
