+++ src/main/java/me/meloni/SolarLogAPI/Interface/BasicUI/MonthCustomizer.java
@@ -1,8 +1,8 @@
 package me.meloni.SolarLogAPI.Interface.BasicUI;
 
 import me.meloni.SolarLogAPI.SolarMap;
-import me.meloni.SolarLogAPI.Interface.DatePicker;
-import me.meloni.SolarLogAPI.Interface.Graph.MonthView;
+import me.meloni.SolarLogAPI.Interface.Components.DatePicker;
+import me.meloni.SolarLogAPI.Interface.Components.Graph.MonthView;
 import me.meloni.SolarLogAPI.DataConversion.GetStartOf;
 
 import javax.swing.*;
