+++ src/main/java/me/meloni/SolarLogAPI/Interface/Components/DatePicker.java
@@ -1,4 +1,4 @@
-package me.meloni.SolarLogAPI.Interface;
+package me.meloni.SolarLogAPI.Interface.Components;
 
 import me.meloni.SolarLogAPI.SolarMap;
 import me.meloni.SolarLogAPI.DataConversion.Entries;
