+++ src/main/java/me/meloni/SolarLogAPI/Interface/Components/Graph/MonthView.java
@@ -1,4 +1,4 @@
-package me.meloni.SolarLogAPI.Interface.Graph;
+package me.meloni.SolarLogAPI.Interface.Components.Graph;
 
 import javax.swing.*;
 import java.awt.*;
